git-svn-id: svn+ssh://erebor.esa.informatik.tu-darmstadt.de/home/wimi/svn/GaalopCompiler/gaalop-2.0/trunk@1271 4235ee25-d3ef-403b-b6ed-692c26853d4b

+++ api/src/main/java/de/gaalop/gapp/variables/GAPPConstant.java
@@ -6,9 +6,9 @@ package de.gaalop.gapp.variables;
  */
 public class GAPPConstant extends GAPPValueHolder {
 
-    private float value;
+    private double value;
 
-    public GAPPConstant(float value) {
+    public GAPPConstant(double value) {
         this.value = value;
     }
 
@@ -17,17 +17,17 @@ public class GAPPConstant extends GAPPValueHolder {
         return false;
     }
 
-    public float getValue() {
+    public double getValue() {
         return value;
     }
 
-    public void setValue(float value) {
+    public void setValue(double value) {
         this.value = value;
     }
 
     @Override
     public String prettyPrint() {
-        return Float.toString(value);
+        return Double.toString(value);
     }
 
     @Override
